requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
crispy-bootstrap4==2022.1
cryptography==44.0.2
distro==1.9.0
dj-database-url==2.2.0
Django==4.2.21
django-crispy-forms==2.0
django-environ==0.11.2
django-webpush==0.3.6
djangorestframework==3.15.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
frozenlist==1.4.1
gunicorn==23.0.0
h11==0.14.0
http_ece==1.2.1
httpcore==1.0.7
httpx==0.28.1
idna==3.6
importlib_metadata==8.5.0
jiter==0.8.2
lxml==5.3.0
multidict==6.0.5
numpy==1.24.4
openai==1.63.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.0.3
pdfkit==1.0.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
py-vapid==1.9.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==11.1.1
PyJWT==2.8.0
PyMuPDF==1.24.11
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
pyppeteer==2.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.2
pywebpush==2.0.3
requests==2.31.0
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
striprtf==0.0.28
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.12.0
twilio==9.0.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
whitenoise==6.7.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.20.2
setuptools>=42
wheel
